Extracted Class MethodAnalyzer
